+++ tests/Command/Response/QscanResponseTest.php
@@ -1,24 +1,24 @@
 <?php
-namespace Disque\Test\Command;
+namespace Disque\Test\Command\Response;
 
 use PHPUnit_Framework_TestCase;
 use Disque\Command\Hello;
 use Disque\Command\Response\ResponseInterface;
-use Disque\Command\Response\CursorResponse;
+use Disque\Command\Response\QscanResponse;
 use Disque\Command\Response\InvalidResponseException;
 
-class CursorResponseTest extends PHPUnit_Framework_TestCase
+class QscanResponseTest extends PHPUnit_Framework_TestCase
 {
     public function testInstance()
     {
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $this->assertInstanceOf(ResponseInterface::class, $r);
     }
 
     public function testInvalidBodyNotArrayString()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: "test"');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody('test');
     }
@@ -26,7 +26,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyNotArrayNumeric()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: 128');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(128);
     }
@@ -34,7 +34,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElementsNotEnough()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: ["10"]');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['10']);
     }
@@ -42,7 +42,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElementsTooMany()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: ["10",["queue1"],"test"]');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['10', ['queue1'], 'test']);
     }
@@ -50,7 +50,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElements0NotSet()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: {"9":"10","1":["queue1"]}');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody([9=>'10', 1=>['queue1']]);
     }
@@ -58,7 +58,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElements1NotSet()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: {"0":"10","2":["queue1"]}');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody([0=>'10', 2=>['queue1']]);
     }
@@ -66,7 +66,7 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElement1NotNumeric()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: ["test",["queue1"]]');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['test', ['queue1']]);
     }
@@ -74,14 +74,14 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
     public function testInvalidBodyElement2NotArray()
     {
         $this->setExpectedException(InvalidResponseException::class, 'Invalid command response. Command Disque\\Command\\Hello got: ["test","queue1"]');
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['test', 'queue1']);
     }
 
-    public function testParseNoQueues()
+    public function testParseNoValues()
     {
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['0', []]);
         $result = $r->parse();
@@ -93,9 +93,9 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
         ], $result);
     }
 
-    public function testParseOneQueue()
+    public function testParseOneValue()
     {
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['0', ['queue1']]);
         $result = $r->parse();
@@ -108,9 +108,9 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
         ], $result);
     }
 
-    public function testParseSeveralQueues()
+    public function testParseSeveralValues()
     {
-        $r = new CursorResponse();
+        $r = new QscanResponse();
         $r->setCommand(new Hello());
         $r->setBody(['1', ['queue1', 'queue2']]);
         $result = $r->parse();
@@ -123,5 +123,4 @@ class CursorResponseTest extends PHPUnit_Framework_TestCase
             ]
         ], $result);
     }
-
 }